m8-task1: Первая партия исправлений по замечаниям.
